# backend/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "DevOps Disaster Simulator API",
    "description": "API del simulador de desastres DevOps. Gestiona sesiones de juego, estado de la simulación, acciones del jugador, diagnósticos SSH, control de la simulación y ranking de partidas.",
    "version": "2.0.0"
  },
  "servers": [
    { "url": "http://localhost:3001", "description": "Backend local" }
  ],
  "components": {
    "securitySchemes": {
      "SessionKey": {
        "type": "apiKey",
        "in": "header",
        "name": "x-session-key",
        "description": "Clave de sesión devuelta por POST /api/session. Requerida en todos los endpoints excepto /api/session y /api/ranking."
      }
    },
    "schemas": {
      "Error401": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": false },
          "message": { "type": "string", "example": "Falta x-session-key header" }
        }
      }
    }
  },
  "security": [{ "SessionKey": [] }],
  "paths": {
    "/api/session": {
      "post": {
        "summary": "Crear una sesión de juego",
        "description": "Crea una nueva sesión e inicia la simulación. Devuelve una sessionKey que debe enviarse como header `x-session-key` en todas las peticiones posteriores.",
        "tags": ["Sesión"],
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["nickname"],
                "properties": {
                  "nickname": {
                    "type": "string",
                    "description": "Nombre del jugador",
                    "example": "xabi_ops"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Sesión creada correctamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "sessionKey": { "type": "string", "description": "Clave de sesión (hex de 32 caracteres)", "example": "a1b2c3d4e5f6a1b2c3d4e5f6a1b2c3d4" },
                    "nickname": { "type": "string", "example": "xabi_ops" }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Nickname vacío o no proporcionado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": false },
                    "message": { "type": "string", "example": "Nickname requerido" }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Cerrar sesión (salida voluntaria)",
        "description": "Cierra la sesión del jugador y marca la partida como terminada con motivo 'quit'. La partida queda registrada en el ranking.",
        "tags": ["Sesión"],
        "responses": {
          "200": {
            "description": "Sesión cerrada correctamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "message": { "type": "string", "example": "Sesión cerrada" }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Sesión no proporcionada o expirada",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error401" }
              }
            }
          }
        }
      }
    },
...
  },
...
}